+++ backend/apps/other/api/urls.py
@@ -6,4 +6,5 @@ urlpatterns = [
     path("categories/", views.CategoryListAPIView.as_view()),
     path("companies/", views.CompanyListCreateAPIView.as_view()),
     path("companies/<int:pk>", views.CompanyDetailAPIView.as_view()),
+    path("skills/", views.SkillsListAPIView.as_view()),
 ]